Refactor osf-model unit test to use mirage
